requirements.txt
# The Vault - Dependencies
# Django 5.x with production stack

# Core Framework
Django>=5.0,<6.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Static Files
whitenoise>=6.6.0

# Media Storage (Cloudinary)
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0

# Production Server
gunicorn>=21.2.0

# Security
python-dotenv>=1.0.0

